Fix type conversion warning

+++ code/31_compute_shader.cpp
@@ -222,7 +222,7 @@ private:
             drawFrame();
             // We want to animate the particle system using the last frames time to get smooth, frame-rate independent animation
             double currentTime = glfwGetTime();
-            lastFrameTime = (currentTime - lastTime) * 1000.0;
+            lastFrameTime = (float)((currentTime - lastTime) * 1000.0);
             lastTime = currentTime;
         }
